miienlp/sentimentAnalysis/requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
exceptiongroup==1.1.2
filelock==3.12.2
fsspec==2023.1.0
huggingface-hub==0.0.8
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
joblib==1.3.1
Levenshtein==0.21.1
numpy==1.21.6
packaging==23.1
pandas==1.3.5
pickle5==0.0.11
pipdeptree==2.9.4
pluggy==1.2.0
pytest==7.4.0
python-dateutil==2.8.2
python-Levenshtein==0.21.1
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.1.2
regex==2023.6.3
requests==2.31.0
sacremoses==0.0.53
safetensors==0.3.1
six==1.16.0
tokenizers==0.10.3
tomli==2.0.1
torch==1.13.1
tqdm==4.65.0
transformers==4.6.1
typing_extensions==4.7.1
urllib3==2.0.3
zipp==3.15.0
